staan nu naast elkaar met border
